remaniement Damage class, ecriture test
