+++ model/ressource_model.py
@@ -1,4 +1,4 @@
-class RessourceModel:
+class RessourceModel(object):
     """
     Contain information for ressources
     constructor contain  argument string
@@ -16,56 +16,47 @@ class RessourceModel:
         return self._name
 
     @name.setter
-    def name(self,name):
+    def name(self, name):
         self._name = name
-
 
     @property
     def vmin(self):
         return self._vmin
 
-
     @vmin.setter
-    def vmin(self,vmin):
+    def vmin(self, vmin):
         self._vmin = vmin
-
 
     @property
     def vmax(self):
         return self._vmax
 
-
     @vmax.setter
-    def vmax(self,vmax):
+    def vmax(self, vmax):
         self._vmax = vmax
-
 
     @property
     def vdefault(self):
         return self.vdefault
 
-
     @vdefault.setter
-    def vdefault(self,vdefault):
+    def vdefault(self, vdefault):
         self.vdefault = vdefault
 
     def get_values(self):
-        return {"vmin":self._vmin, "vmax":self._vmax, "vact":self._vact, "vdefault":self._vdefault}
+        return {"vmin": self._vmin, "vmax": self._vmax, "vact": self._vact, "vdefault": self._vdefault}
 
-    def set_values(self, dictionnary_of_values):
-        name_of_values = ["vmin", "vmax", "vact", "vdefault" ]
+    def set_values(self, dictionnary_of_values: dict):
+        name_of_values = ["vmin", "vmax", "vact", "vdefault"]
         try:
             for key in dictionnary_of_values.keys():
-                if not key in name_of_values :
+                if not key in name_of_values:
                     return False
-        except AttributeError :
+        except AttributeError:
             raise AttributeError("Le paramètre doit être un dictionnaire")
         else:
             self._vact = dictionnary_of_values['vact']
             self._vmin = dictionnary_of_values['vmin']
             self._vmax = dictionnary_of_values['vmax']
             self._vdefault = dictionnary_of_values['vdefault']
-
-
-
-
+            return True
